requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
openai==1.6.0 
greenlet==2.0.2
